Listing: database refactoring
Removes craft_id because this will be a column of Job model instead. Removes is_offering because Listing will be renamed to JobVacancy instead and hence 'listing' is now understood as 'a vacancy to perform a role on a project'. Renames 'user_id' to 'posted_by'. Adds the following columns: 'slug', 'is_active', 'ends_at', 'deleted_at'