Creats Artisan command for exporting `crafts` table as a nested JSON array.
With the generated JSON file, the server will not have to query the database to figure out the hierarchy of the categories. The output JSON file can be used by the Frontend Developer to build the marketplace app's subnavigation. In the future, we might want to consider adding an API endpoint to this JSON data.