API Design
These last 3 days were dedicated to designing the RESTful API architecture. A .xlsx spreadsheet is being maintained for this purpose. This file is external to our git repo since git cannot handle this file format. Instead a .csv export has been added and will be updated with a new export anytime significant changes are made to the .xlsx file.